Port over tests from koji_wrapper.
This patch ports some more extensive testing added when these functions
lived in koji_wrapper, since that project now depnds on this one.  Also,
it puts them in the tests/unit/ directory, which a subsequent patch with
do with the remaining unit tests.  This lays the groundwork for
supporting other types of tests if we find they are needed in the
future, and brings toolchest more in line with the structure of our
other projects.

Signed-off-by: Jason Guiditta <jason.guiditta@gmail.com>