cpaas_label_compare: Update for new CPaaS rules
Previously, CPaaS was updating the release number at the
front of the release. Now, it's appending a dot and a rebuild
count number after the dot (and possibly incrementing it).

So, update the comparison function to chop off the last dot - and
following number (if it exists) and compare using the truncated
version.

Signed-off-by: Lon Hohberger <lhh@redhat.com>