Deprecate dlrn_label_compare & stop wrapping it
1. dlrn_label_compare is an artifact from older releases where we
would switch to non-DLRN builds just prior to releasing software.

For example, we might have '1.0.1-0.20190403103245.a3fc4ee' and
then, right before release, we would switch to the '1.0.0-1.rc1'
or '1.0.0-1' build instead, which we wanted to supersede the DLRN
build.  We stopped doing this some time ago, preferring to simply
ship the content as it was built by DLRN rather than attempting
to change it last-minute.

Since there is no known current release doing this, stop using it
by default and deprecate it, since it creates unexpected results
and is DLRN-deployment specific.

2. label_compare (labelCompare) is largely a copy of C algorithms
from RPM's library, so we want it to behave like librpm, not wrap
another function.

Signed-off-by: Lon Hohberger <lhh@redhat.com>