Fix documentation on rtd.
Previously, we were not generating the actual api docs,
only the included files from above docs/ dir.

NOTE: Docs for our yaml module currently do not build, due
to a naming conflict with pyyaml. Sphinx thinks there is a
recursive dependency currently.  This will be addressed in
a future patch.

Signed-off-by: Jason Guiditta <jguiditt@redhat.com>