Add auto-publishing and changelog generation.
This was already added to most of our other projects, but toolchest did
not yet have this tooling.  Each commit should be published to
test-pypi, and on tag publish to pypi. Also generated the changelog as
we do elsewhere.  Patch includes a couple minor cleanups of related
tooling for consistency as well.

Signed-off-by: Jason Guiditta <jguiditt@redhat.com>